# Tagged enum usability & performance improvements