--- a/rustler_tests/native/rustler_test/src/test_codegen.rs
+++ b/rustler_tests/native/rustler_test/src/test_codegen.rs
@@ -92,7 +92,7 @@ pub enum TaggedEnum1 {
     String1(String),
     String2(String),
     Untagged,
-    Samefields {x: i32, y: i32},
+    Samefields { x: i32, y: i32 },
 }
 
 #[rustler::nif]