1. Support multiple units with same field names
2. Move atom / tuple decode outside the type check if, to remove duplicate decode calls

Enum with 22 cases:
before:
tagged_enum_decode          1.56 M      642.48 ns  ±1292.75%         634 ns         675 ns
after:
tagged_enum_decode          3.17 M      314.98 ns  ±2092.02%         311 ns         343 ns

(~300ns is also the tagged_enum_encode latency on my machine)